refactor(schemas): single-source MCP tool input schemas into toolSchemas.js
All 25 tool {description, inputSchema} definitions (the zod shapes) plus the
confirmation tokens and the XAE_ACTIONS map now live once in toolSchemas.js;
index.js registers each tool via a lookup. This removes the schema half of the
dual-maintenance that previously required editing index.js in two places to add
a parameter (e.g. search's recent `refresh`).

Handlers and their input->payload transforms (field renames like path->treePath,
save coercions, confirm-gates, the xae action-enum mapping) are intentionally
left in the handlers, since those are real logic, not schema. The C# daemon is
untouched and was never restarted.

Verified schema-preserving: tools/list captured over stdio before and after is
byte-for-byte identical across all 25 tools (names, descriptions, inputSchemas).
node --check passes; a live round-trip (tc_system get_netid, plc_pou search) was
confirmed against the running warm daemon.
